requirements.txt
django
djangorestframework
psycopg2
django-cors-headers
pillow
django-rest-auth
djangorestframework-jwt